+++ src/Testing/HandlesAuthentication.php
@@ -6,8 +6,6 @@ use BristolSU\ControlDB\Contracts\Models\Group;
 use BristolSU\ControlDB\Contracts\Models\Role;
 use BristolSU\ControlDB\Contracts\Models\User;
 use BristolSU\Support\Authentication\Contracts\Authentication;
-use BristolSU\Support\Permissions\Contracts\PermissionTester;
-use Prophecy\Argument;
 
 /**
  * Trait for aiding interactions with the user/group/role system
@@ -84,48 +82,5 @@ trait HandlesAuthentication
     {
         app()->make(Authentication::class)->setUser($user);
     }
-
-    /**
-     * Bypass any authorization checks made in controllers.
-     * 
-     * This will simply ensure any call to the permission tester 'evaluate' method returns true.
-     * 
-     * @return void
-     */
-    public function bypassAuthorization()
-    {
-        $permissionTester = $this->prophesize(PermissionTester::class);
-        $permissionTester->evaluate(Argument::any())->willReturn(true);
-        $this->instance(PermissionTester::class, $permissionTester->reveal());
-    }
-
-
-
-//    /**
-//     * @param $method
-//     * @param $route
-//     * @param null $ability
-//     * @param array $parameters
-//     */
-//    public function assertRequiresAuthorization($method, $route, $ability = null, $parameters = [])
-//    {
-//        $this->beUser($this->user);
-//        $this->beGroup($this->group);
-//        $this->beRole($this->role);
-//        $this->be($this->databaseUser, 'web');
-//        $this->be($this->databaseUser, 'api');
-//        $response = $this->call($method, $route, $parameters);
-//        $response->assertStatus(403, 'User allowed past authorization without permission. Is there an \'authorize\' statement?');
-//
-//        $permissionTester = $this->prophesize(PermissionTester::class);
-//        $permissionTester->evaluate($this->alias() . '.' . $ability)->shouldBeCalled()->willReturn(true);
-//        $this->instance(PermissionTester::class, $permissionTester->reveal());
-//
-//        $response = $this->call($method, $route, $parameters);
-//        $this->assertTrue(
-//            $response->isSuccessful(),
-//            sprintf('User not allowed past authorization with permission. Status code %s', $response->getStatusCode())
-//        );
-//
-//    }
+    
 }